Mobile-Devices-Scrapper/requirements.txt
beautifulsoup4
requests
Flask
loguru
lxml
mysql-connector-python
